pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "script-builder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.54.0",
    "exa-py>=1.16.1",
    "google-genai>=1.44.0",
    "python-dotenv>=1.1.0",
    "xmltodict>=0.14.2",
]

[tool.hatch.build.targets.wheel]
packages = ["utils"]
